+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Patch Changes
 
-- Enable child node loading and fold/unfold via keyboard without requiring an initial mouse click.
+- a11y : improves keyboard accessibility in the TreeView component
+- a11y : Enable child node loading and fold/unfold via keyboard without requiring an initial mouse click.
 
 ## 0.16.1